--- a/addon/core/__init__.py
+++ b/addon/core/__init__.py
@@ -1,0 +1,9 @@
+"""
+Core package for Blender Driving Scenario Creator.
+
+This package contains the fundamental classes and utilities used throughout the add-on.
+"""
+
+from .constants import *
+from .exceptions import *
+from .config import *