--- a/addon/road_objects/__init__.py
+++ b/addon/road_objects/__init__.py
@@ -1,0 +1,20 @@
+"""
+Road Objects package for the Driving Scenario Creator add-on.
+
+This package contains all road object-related classes including signs,
+traffic lights, stop lines, and stencils.
+"""
+
+from .road_object_base import RoadObjectBase
+from .road_object_sign import RoadObjectSign
+from .road_object_traffic_light import RoadObjectTrafficLight
+from .road_object_stop_line import RoadObjectStopLine
+from .road_object_stencil import RoadObjectStencil
+
+__all__ = [
+    'RoadObjectBase',
+    'RoadObjectSign',
+    'RoadObjectTrafficLight', 
+    'RoadObjectStopLine',
+    'RoadObjectStencil'
+]